Move to lib/drivers and test
