+++ robot/src/main/java/us/ilite/lib/drivers/GetLocalIP.java
@@ -1,4 +1,4 @@
-package us.ilite.robot.os.commands;
+package us.ilite.lib.drivers;
 
 import java.io.BufferedReader;
 import java.io.IOException;
